Timer and keyup/keydown rework
Bindings are now stored in a table private to the input module; _G.input
is the public API only. bind(), unbind() and unbindall() are used to
manipulate this table. Keydown generates a "key_X" event, keyup generates
a "!key_X" event - this isn't supported for joysticks yet. Key repeat is
enabled via input.setRepeat(event, delay, period), and works on *any* event
including virtual events like "up" and "down" - not just key events.

input.bind can be used in a few ways. input.bind("foo", "bar", "baz") binds the
events "foo" and "!foo" to the events "bar" and "baz". bind("foo", fnA, fnB)
binds the event "foo" to fnA and "!foo" to fnB. If the second function/event
to bind to is omitted, !foo is bound to !bar in the first case and to nil
in the second.

These bindings overwrite any existing bindings for that event.