Replace Class:new with new "Class", use property tables for Node constructor
Riding the refactor tractor for a bit. Node and all of its subclasses now take a property
table that is shallow-copied into the instance. This makes it easier to create Nodes
with arbitrary properties without needing to keep the instance around and set properties
on it by hand. Some old uses of calling the :new method directly have also been replaced
with calls to the global new() function.