Fix error where going up from the top of the media library would crash emufun.
Specifically, because of the way the Configuration Files vdir was
constructed, at __init time it had an invalid path() method. This caused
the cache information lookup to fail (since it couldn't be statted) and
this in turn meant that when it was time to draw it, it had no cache
information with which to determine what colour to draw it in.